test(datetime-button): migrate overlays
